requirements.txt
streamlit
pandas
openpyxl
matplotlib
google-generativeai